suppress update notfications in Wayland/Wayfire

+++ install.sh
@@ -46,11 +46,16 @@ ufw enable
 systemctl enable ssh
 systemctl start ssh
 
-# suppress update notfications in X11/LXDE
-if [ ! -f /home/$SUDO_USER/.config/lxpanel/LXDE-pi/panels/panel ]; then
+# suppress OS update notifications
+if [ "$XDG_SESSION_TYPE" == "wayland" ]; then
+  apt-get -y install crudini
+  crudini --inplace --set --ini-options=nospace ~/.config/wf-panel-pi.ini panel updater_interval 0
+else
+  if [ ! -f /home/$SUDO_USER/.config/lxpanel/LXDE-pi/panels/panel ]; then
     mkdir /home/$SUDO_USER/.config/lxpanel/LXDE-pi/panels
     cp /etc/xdg/lxpanel/LXDE-pi/panels/panel /home/$SUDO_USER/.config/lxpanel/LXDE-pi/panels/panel
     chown -R $SUDO_USER /home/$SUDO_USER/.config/lxpanel/
+  fi
+  sed -i -E '/  type=updater/{N;N;N;s/  Config \{\n(.*\n)?  \}/  Config \{\n    Interval=0\n  \}/}' /home/$SUDO_USER/.config/lxpanel/LXDE-pi/panels/panel
+  killall -SIGHUP lxpanel
 fi
-sed -i -E '/  type=updater/{N;N;N;s/  Config \{\n(.*\n)?  \}/  Config \{\n    Interval=0\n  \}/}' /home/$SUDO_USER/.config/lxpanel/LXDE-pi/panels/panel
-killall -SIGHUP lxpanel
